+++ NEWS.md
@@ -1,7 +1,14 @@
+# eiCompare 3.0.5
+
+## added function 11/11/25
+
+* included extract_rxc_precinct() function to extract precinct level estimates from ei_rxc()
+
 # eiCompare 3.0.4
 
 ## Package changes
-
+* added rpv_normalize() function
+* removed wru dependency
 * incorporated rpv_coef_plot() and rpv_toDF() functions from eiExpand package
 * edited ei_iter() to have flexible CI parameters (default is 0.95) using bayestestR for calculation and updated column naming, and to use reproducible parallel processing (.inorder=TRUE)
 * edited ei_rxc() with repdocuible parallel processing and changed column naming to fit ei_iter()